add idea of scene
