+++ index.js
@@ -1,3 +1,26 @@
 
-exports.node = require('./lib/node');
-exports.component = require('./lib/component');
+/**
+ * Module dependencies.
+ */
+
+var component = require('./lib/component');
+var scene = require('./renderer/scene');
+var node = require('./lib/node');
+
+/**
+ * Expose `Scene`.
+ */
+
+exports.scene = scene;
+
+/**
+ * Expose `component`.
+ */
+
+exports.component = component;
+
+/**
+ * Expose `node`.
+ */
+
+exports.node = node;
