+++ lib/node/index.js
@@ -3,9 +3,9 @@
  * Module dependencies.
  */
 
+var ComponentNode = require('./component');
 var ElementNode = require('./element');
 var TextNode = require('./text');
-var ComponentNode = require('./component');
 
 /**
  * Expose `node`.
