+++ lib/renderer/component.js
@@ -1,4 +1,9 @@
-var Tree = require('./tree');
+
+/**
+ * Module dependencies.
+ */
+
+var Tree = require('../tree');
 var i = 0;
 
 /**
@@ -6,10 +11,10 @@ var i = 0;
  * all of it's components also rendered into sub-trees
  *
  * @param {Component} component
- * @param {RenderedComponent} owner
+ * @param {ComponentRenderer} owner
  */
 
-function RenderedComponent(component, owner) {
+function ComponentRenderer(component, owner) {
   this.id = (++i).toString(36);
   this.root = this.owner ? this.owner.root : this;
   this.component = component;
@@ -22,10 +27,10 @@ function RenderedComponent(component, owner) {
 /**
  * Render the tree and it's sub-trees
  *
- * @param {RenderedComponent} rendered
+ * @param {ComponentRenderer} rendered
  */
 
-RenderedComponent.prototype.render = function() {
+ComponentRenderer.prototype.render = function(){
   var self = this;
   var top = this.tree.root;
   var components = this.components;
@@ -33,11 +38,11 @@ RenderedComponent.prototype.render = function() {
   this.tree.walk(function(path, node){
     if (node.type === 'component') {
       var instance = node.create();
-      var rendered = new RenderedComponent(instance, self);
+      var rendered = new ComponentRenderer(instance, self);
       components[path] = rendered;
     }
   });
-}
+};
 
 /**
  * Diff this tree with another Node
@@ -47,10 +52,10 @@ RenderedComponent.prototype.render = function() {
  * @return {Path}
  */
 
-RenderedComponent.prototype.diff = function(target) {
+ComponentRenderer.prototype.diff = function(target){
   var tree = this.tree;
 
-  function walk(node){
+  function walk(node) {
     var path = tree.path(node);
     var matching = target.get(path);
     fn(path, node, matching);
@@ -60,4 +65,4 @@ RenderedComponent.prototype.diff = function(target) {
   }
 
   walk(this.root);
-}
+};
