+++ lib/tree/diff.js
