+++ lib/tree/patch.js
